Создал модульную структуру js-файлов
